Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> Tom, are you aware of any problems (crashing) with init in AOSP master?

I'm not aware of any.  Is this using a plain ramdisk to boot (with a statically linked first stage init)?  There were some recent changes with libcrypto to have a self test during that linking that may be involved.  Other than that, are we sure that the ramdisk is getting loaded and init is found and loaded?

Init (1st and 2nd stage) creates signal handlers very very early that print backtraces when it crashes now, so this failure must happen very early.

Patch-set: 8
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>