pixel3_maineline: Add changes to get pixel3_mainline building a boot.img
This adds some changes to get the pixel3_mainline target
building a boot.img. It requires that a pixel3_mainline dir be
present in the dragonboard-kernel dir, and it should contain
a Image.gz-dtb file and any kernel modules to be included.

If you do not want to use kernel modules, build with
PIXEL3_USES_GKI=false.

Change-Id: Iacea2afd0c81d65935de7e52bb680d2e314f6b3c
Signed-off-by: John Stultz <john.stultz@linaro.org>