pixel3_maineline: Add changes to get pixel3_mainline building a boot.img (v2)
This adds some changes to get the pixel3_mainline target
building a boot.img. It requires that a pixel3_mainline dir be
present in the dragonboard-kernel dir, and it should contain
a Image.gz-dtb file and any kernel modules to be included.

If you do not want to use kernel modules, build with
PIXEL3_USES_GKI=false.

This is take number two, as the first pass was causing build
breakage. In this version, I check to make sure we have a
kernel image before we try to build

Change-Id: I5810ee98ce313521bbe56831cb298b55674530fc
Signed-off-by: John Stultz <john.stultz@linaro.org>