pixel3_mainline: Add changes to get pixel3_mainline building a boot.img (v3)
This adds some changes to get the pixel3_mainline target
building a boot.img. It requires that a pixel3_mainline dir be
present in the dragonboard-kernel dir, and it should contain
a Image.gz-dtb file and any kernel modules to be included.

If you do not want to use kernel modules, build with
PIXEL3_USES_GKI=false.

This is take number two, as the first pass was causing build
breakage. In this version, I check to make sure we have a
kernel image before we try to build.

In v3, I've moved some more modules as vendor-only, to allow
the pixel3 to boot.

Signed-off-by: John Stultz <john.stultz@linaro.org>
 [sumits: updated the list of vendor-only modules]
Signed-off-by: Sumit Semwal <sumit.semwal@linaro.org>
Change-Id: Ie94ef4af00c415baeb20838cb0ba30434efa691e