db845c: Wire in bluetooth support
The qca bluetooth module has some yet to be determined issue
which causes deferred firmware loading to stop it from working.

But as long as the bluetooth module is loaded from the vendor
partition where the bluetooth firmware is present, it can be
made to load and function.

This patch updates the bluetooth hal version and adds a trick
to exclude the qca and hci_uart modules from being on the
ramdisk.

Note: Bluetooth audio isn't working yet. Devices pair but I
don't hear audio yet.

Change-Id: I37e927485414368b817958dfef2137fc263003a0
Signed-off-by: John Stultz <john.stultz@linaro.org>