db845c: firmware: Use qcom/sdm845/ prefix on gpu firmware files
This copies the gpu firmware files to be under firmware/qcom/sdm845/
since that is where they will be expected to be with newer kernels.

One note: v5.6+ kernels not only expect a different path but also
expect a different filename postfix. So this copies the a630_zap.mdt
to a630_zap.mbn.

For compatiblility with older kernels, we keep the existing firmware
as is.

Change-Id: I8e81360e573c68e912378a3493c9ef9104f23dc6
Signed-off-by: John Stultz <john.stultz@lianro.org>