WIP: db845c: firmware: Use qcom/sdm845/ prefix on gpu firmware files
This moves the gpu firmware files to be under firmware/qcom/sdm845/
since that is where they will be expected to be with newer kernels.

For compatiblility with older kerenls, duplicate the zap firmware
file, preserving the old location.

NOTE: This still has issues as upstream kernel wants a630_zap.mbn
but we only have a630_zap.mdt.  Will update this once that gets
sorted out

Change-Id: I8e81360e573c68e912378a3493c9ef9104f23dc6
Signed-off-by: John Stultz <john.stultz@lianro.org>