db845c: firmware: Use qcom/sdm845/ prefix on gpu firmware files
This copies the gpu firmware files to be under firmware/qcom/sdm845/
since that is where they will be expected to be with newer kernels.

One note: v5.6+ kernels not only expect a different path but also
expect a different filename postfix. So this copies the a630_zap.mdt
to a630_zap.mbn.

For compatiblility with older kernels, we keep the existing firmware
as is.

Change-Id: I3ae4706f9a8ef36c72a98e34ae50cac562071d86
Signed-off-by: John Stultz <john.stultz@lianro.org>