sepolicy: Allow access to graphics_device getattr
dmesg and logcat is filled with selinux denials
related to getting access to graphics device
attributes.

From bootanimation to all the way up to video
playback, all display/graphics related apps/services
seem to need access to graphics device attributes.
Although the denials are non-fatal but it is good
to get rid of the denial messages.

Change-Id: Ic65364c08b5de8aab8b22997ab5215304a4782ac
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>