db845c: Make userdata formattable
If one runs "fastboot erase userdata", and then does not flash
a userdata image to the partition, the system will fail to boot
as /data/ will not mount, and the apex and dalvik cache
directories won't be present or writable.

This tries to address this by adding the formattable flag to
the fstab and add the partition to the sepolicy so mkfs can
access it.

NOTE: I did have to use the full by-name path in the fstab,
which is less flexible then the LABEL= method, but if the
partition isn't formatted there isn't neceessarily a partition
label to find.

Big thanks to AmitP for suggesting this approach

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Ia678d0d1d5e1eb35462f8fcfee478969c614b0d4