db845c: Install BT driver modules to ramdisk
With following android-5.4 kernel changes, we no
longer need to install and load the bt driver
modules from the vendor partition:
https://android-review.googlesource.com/c/1261169
https://android-review.googlesource.com/c/1261170

Also BOARD_VENDOR_KERNEL_MODULES duplicate
installed modules in ramdisk and vendor partition,
so remove that part. And rename
BOARD_VENDOR_RAMDISK_KERNEL_MODULES to
BOARD_GENERIC_RAMDISK_KERNEL_MODULES to avoid
confusion because we are not supporting vendor
boot image at the moment.

With this change all the kernel modules are now
installed in ramdisk.img, and we don't need to
update and flash vendor.img to test kernel changes.

Change-Id: Ieccce38e4eb3f915e75a2c9b636b4a2d7bca586b
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>