Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I think the sysfs file name change was an upstream kernel change. So I guess this change will break older kernels?

I tested with the prebuilt kernels, it works.

If my understanding is right, that the name of soc part is decided by the definition in DTS like here:
https://android.googlesource.com/kernel/common/+blame/refs/heads/android-mainline/arch/arm64/boot/dts/qcom/sdm845.dtsi#705

which was updated at about July 25th, 2019 with the commit here:
https://android.googlesource.com/kernel/common/+/a1875bf98290e538de2484c4db957ace5a8ac578%5E%21/#F0

Then your guess is right, it would break kernels older than July 25th, 2019.

Patch-set: 2