init.common.rc: Set /sys/kernel/debug/tracing permissions
CONFIG_DEBUG_FS is enabled in android-mainline kernel,
https://android-review.googlesource.com/c/1324220,
but /sys/kernel/debug/tracing permissions are not getting
set in system/core/rootdir/init.rc's existing "on late-fs"
trigger, which resulted in webview crash on db845c.

    F webview_zygote: jni_internal.cc:649] JNI FatalError called: (zygote) \
    Failed open(/sys/kernel/debug/tracing/trace_marker, 1): Permission denied
    ----->8-----snip-----8<-----
    F DEBUG   : signal 6 (SIGABRT), code -1 (SI_QUEUE), fault addr --------
    F DEBUG   : Abort message: 'JNI FatalError called: (zygote) Failed
                open(/sys/kernel/debug/tracing/trace_marker, 1): Permission denied'

This webview crash can be reproduced with stock AOSP
webview version "74.0.3729.186", as well as version
"85.0.4150.0" built locally.

Here is a temporary workaround to set the permissions
on boot completion, while we debug it further.

Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: I177cf8e833d9548a11363a3b89bbdae9c594039a