Update patch set 1
Patch Set 1:

Also I don't who is the right guy to tag for fixing it in the system/core project, so if you guys can forward this issue to them.

I could only reproduce this issue on db845c and pocof1, but I'm sure it must be biting other platforms too.

Patch-set: 1