WIP: db845c: Switch gralloc to use gralloc.minigbm_msm instead of gralloc_gbm
Instead of carying our own copy of gralloc_gbm, try to utilize
the minigbm gralloc as it implements gralloc4

NOTE: Unfortunately, recent changes in minigbm have caused this to
break. Its not clear whats going wrong yet, and it may be
misconfiguration here or something wrong in mesa or drm_hwcomposer
(or something else wrong in minigbm).

So have to hold on this until that's sorted out.

Change-Id: I628ff76d3ef7cf441b122bc6f2e49f662b0c7ed5
Signed-off-by: John Stultz <john.stultz@linaro.org>