WIP: db845c: Switch gralloc to use gralloc.minigbm_msm instead of gralloc_gbm
Instead of carying our own copy of gralloc_gbm, try to utilize
the minigbm gralloc as it implements gralloc4

NOTE #1: This depends on a change to external/minigbm to enable
msm support

NOTE #2: For this change to work properly, it seems to have a
dependency on an updated mesa. This is fairly complex, as the reason
seems to be that drm_hwcomposer in AOSP has changes for minigbm to
support format_modifiers that landed in a later mesa version.
Unfortunately using them with the current aosp  mesa version causes
strange zebra stripe artifacts on the display.

Thus the ordering for this probably needs to be:
1) external/minigbm changes for -DDRV_MSM
2) New mesa update (may break yuv video in some cases, due to
   existing gbm_gralloc mishandling yuv)
3) This patch to switch db845c to minigbm

Change-Id: I628ff76d3ef7cf441b122bc6f2e49f662b0c7ed5
Signed-off-by: John Stultz <john.stultz@linaro.org>