WIP: db845c: Switch gralloc to use gralloc.minigbm instead of gralloc_gbm
Instead of carying our own copy of gralloc_gbm, try to utilize
the minigbm gralloc as it implements gralloc4

NOTE: For this change to work properly, it seems to have a dependency
on an updated mesa. This is fairly complex, as the reason seems to
be that drm_hwcomposer in AOSP has changes for minigbm to support
format_modifiers that landed in a later mesa version. Unfortunately
using them with a newer mesa version causes strange zebra stripe
artifacts.

NOTE: This requires a hack to external/minigbm

Change-Id: I628ff76d3ef7cf441b122bc6f2e49f662b0c7ed5
Signed-off-by: John Stultz <john.stultz@linaro.org>