db845c: Add missing etc/permissions/android.software.xml
Since Android11 merged in, I have been seeing some odd crashes
occasionally that was pointing to the fact that the
AutoFillManager mService value was null.

However, this same issue was not seen on HiKey960, with the same
application and AOSP source.

I dug around and realized we were not adding the
android.software.xml file we add on HiKey960, and copying it over
resolves the issue.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I1bc88c67030b0ed69511acdf348c27367bf7213a