WIP: db845c: gralloc_gbm: Fix up YV12 plane count
With newer mesa, the code expects YV12 to have three
planes, and with YUV buffers in chrome/webview we
see failures creating buffers here:

https://github.com/pundiramit/external-mesa3d/blob/wip/aosp-merge-mesa-20.3.4//src/gallium/frontends/dri/dri2.c#L956

This patch tries to address this by set all the planes
to the same fd and setting num_fds to the expected 3.

NOTE: This needs more work, and while it fixes the
chrome regression w/ newer mesa, this change breaks
video from the gallery app.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Ice0debaceca7cc3711717d1c69105dbeaec71489