db845c.mk: Do not copy kernel modules to vendor
Keep the kernel modules in ramdisk only for now.

Right now we keep a copy of all the kernel modules
in /vendor/lib/modules, which results in modprobing
of modules from /vendor/lib/modules as well and
though they fail safely but we have run into a
regression where dual loading of UFS module resulted
in probing failures. Ideally dual loading should not
reach up to probing stage. And since this issues is
difficult to reproduce and debug, the easiest way out
is to make sure we do not try to load a kernel module
more than once.

Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: I1be3bccd1ababaf258f1f87c75636919b962ad8d