WIP: db845c: Enabling usb camera
Following:
  https://source.android.com/devices/camera/external-usb-cameras

With a number of additions (like sepolicy, and permissions), in order
to get it going.

This patch also adds a very simple media_profiles.xml implementation
which works for video recording, but adds an iffy vga (640x480) entry
calling it 480p. We will likely need to tweak all this quite a bit
when we add real hwcodec support or onboard camera support.

Change-Id: I9db300d9eadd9246d5f1fbab40d1fcbe33336961
Signed-off-by: John Stultz <john.stultz@linaro.org>