WIP: db845c: Enabling usb camera
Following:
  https://source.android.com/devices/camera/external-usb-cameras

With a number of additions (like sepolicy, and permissions), in order
to get it going.

Currently camera image capture works ok, but video capture crashes
(but video preview does work).

Additionally this patch directly adds the cuttlefish media_profiles.xml
file, so we probably need to tune that file a bit, but there's not much
documentation on the format.

TODO:
* When usb webcam is plugged in, the audio devices are reordered, and
  the audio hal gets confused and output doesn't function.

* Seeing (old) chrome apk crash with:
    System.err: java.lang.IllegalArgumentException: supportsCameraApi:2365: Unknown camera ID 0

Change-Id: I9db300d9eadd9246d5f1fbab40d1fcbe33336961
Signed-off-by: John Stultz <john.stultz@linaro.org>