WIP: db845c: Enabling usb camera
Following:
  https://source.android.com/devices/camera/external-usb-cameras

With a number of additions (like sepolicy, and permissions), in order
to get it going.

Currently camera image capture works ok, but video capture crashes
(but video preview does work).

Additionally this patch directly adds the cuttlefish media_profiles.xml
file, so we probably need to tune that file a bit, but there's not much
documentation on the format.

TODO:
* Video recording still doesn't function properly
* Also web video chat (meeting.jit.si) doesn't properly send/recieve video/audio streams
  - Though it does show local video, but no external

Change-Id: I9db300d9eadd9246d5f1fbab40d1fcbe33336961
Signed-off-by: John Stultz <john.stultz@linaro.org>