HACK: WIP: Cut out 720p recording modes to get webcam video recording working
Managed to chase the video recording failures down to 720p being
an unsupported resolution (i'm guessing for the swdecoder? or maybe
we are missing something else in the config).

Dropping this allows very low res video to be recorded (with audio).

The Jitsi meet app works to record and send video properly. Though
I'm still seeing issues with it not recording audio. The webview
jitsi site still doesn't seem to send/receieve any video or audio.

Need to clean this up & fold down.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I5345d74eae66b22259e2a2c62dd90d18007ff54e