Create a mini configuration for the dragonboards.
The goal of the mini configuration is to define a device configuration
with minimal driver dependencies. The only strict requirements
are storage (UFS) and USB which provides an ADB connection. This
also implies software crypto for userdata because drivers for the
hardware crypto engine are not necessarily available. This allows for
development of low level pieces of the stack such as the kernel and
libc on hardware which does not have a full set of drivers available
in the mainline kernel.

Currently this implies no zygote because of a lack of support for
headless mode. This limits the amount of testing possible but if
this becomes possible in the future we could add the zygote to the
mini configuration.

To show what this would look like for existing devices, add db845c_mini
and rb5_mini targets.

Bug: 217455793
Tested: db845c and db845c_mini boot tested, rb5 and rb5_mini build tested
Change-Id: I8f8b467a4dfad68f6e2e53954b5944c6155de782