dragonboard: fstab.common: Drop inlineencryption to fix adb remount
I recently noticed adb remount was failing on db845c.

After comparing with cuttlefish I narrowed down the issue to the
inlinecrypt option along with the v2+inlinecrypt_optimized
fileencryption flag in the fstab

I'm not sure if this is due to actual breakage or just due to
userland expecting something else these days.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I74fdcb8b7154f37018251fc7b0e6637aacce9570