Enable serial console on dragonboard devices.
A recent android-mainline kernel commit 01e249da0a52 ("ANDROID: GKI:
Enable CONFIG_NULL_TTY by default") enabled CONFIG_NULL_TTY in the
GKI, which caused the serial console to be disabled by default. This
is a reasonable default for production devices, but for development
boards it makes more sense to have it be enabled. Do so by passing
the appropriate console flags in the boot image.

Change-Id: I21e2a53a4f46a698060297dbff717a75816117df