Revert "Remove 64-bit binder ABI setting."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e6cca5f7-f426-4a21-9355-639cbe7dd9c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icba0a9a67818a1d6285810d678290605cbf6489e