Rename qcom folder to utils
We have used qcom/ folder to host various services, scripts
and utilites, and naming the folder to qcom/, to begin with,
was very shortsighted from my end.

So we are renaming it to utils/ and moving it under a new
shared/ folder which should hopefully be our go to
destination for similar common tools/HALs going forward.

Change-Id: I15fe45dbe5bc4752d6a2ee87a3fbed4924a66f08
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>