db845c-userdebug: Enable thermal HAL
This change enables generic thermal HAL on db845c-userdebug
build targets.

The 'etc' contains the thermal configuration for DB845c (aka RB3).

The thermal monitoring daemon is expecting to get a SKIN temperature
sensor. As the RB3/RB5 are evaluation platform there is no such a
sensor on the board so arbitrarly the configuration says the SKIN
sensor is the cluster0. But as this one has fast temperature
transitions, a more convenient sensor may be used later (temperature
speed ~1°C / sec)

Smoke tested on RB5 as well with no obvious regressions. But we
must create a similar configuration file for RB5 as well. Assuming
it will be very close to RB3, that should be trivial to create it.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
[AmitP: Added check to disable this HAL for v5.4 kernel version]
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: I688e7e0db30c8a848f75b7f779bb6fd230c592ce