shared/hal/thermal: Add a generic thermal HAL
This change brings the generic thermal HAL.

These are coming from different git trees merged into this single change.
 - https://git.linaro.org/power/android-power-thermal.git

   It creates device/linaro/dragonboard/shared with the following
   hierarchy:

   shared
   |
    `- hal
       |
       `- thermal
          |
          |-- hal
          |
          |-- lib
          |
           `- sepolicy

The thermal HAL is designed to be as generic as possible in order to
be reused by any platforms in the future.

The in-kernel thermal framework provides more events than what Android
supports like thermal zone creation or deletion. However in the
future, those may be supported, so the thermal HAL has the stubs to be
filled with the logic.

The code is documented, anyone interested in using the generic thermal
HAL can easily fetch the code and understand the purpose of each
function. Nevertheless, the thermal HAL being designed to be generic,
the changes should end up in the common trunk.

[ DanielL:
  - Split the libthermal and the thermal HAL into two different
    patches
  - Fixed Lint warnings]
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
[AmitP: Restructured the original patch and fixed selinux policies]
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: I2594f315fbe283a70443e7ab7150b68496d26236