Add a generic thermal HAL
This change brings the generic thermal HAL.

These are coming from different git trees merged into this single change.
 - https://git.linaro.org/power/libthermal.git
 - https://git.linaro.org/power/android-power-thermal.git

   It creates device/linaro/dragonboard/shared with the following
   hierarchy:

   shared
   |
   `- thermal
      |
      `- etc
      |
      `- hal
      |
      `- lib
      |
      `- sepolicy

The 'etc' contains the thermal configuration for DB845c (aka RB3).
The port to RB5 must create a similar configuration file. Assuming it
will be very close to RB3, that should be trivial to create it.

The thermal HAL is designed to be as generic as possible in order to
be reused by any platforms in the future.

The thermal monitoring daemon is expecting to get a SKIN temperature
sensor. As the RB3/RB5 are evaluation platform there is no such a
sensor on the board so arbitrarly the configuration says the SKIN
sensor is the cluster0. But as this one has fast temperature
transitions, a more convenient sensor may be used later (temperature
speed ~1°C / sec)

The in-kernel thermal framework provides more events than what Android
supports like thermal zone creation or deletion. However in the
future, those may be supported, so the thermal HAL has the stubs to be
filled with the logic.

The code is documented, anyone interested in using the generic thermal
HAL can easily fetch the code and understand the purpose of each
function. Nevertheless, the thermal HAL being designed to be generic,
the changes should end up in the common trunk.

Change-Id: I12aa7719248b2feb36fce9488c560557c5323c89
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
[AmitP: Restructured the original patch, fixed selinux policies and
        added check to disable this HAL for v5.4 kernel version.
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>