shared/hal/thermal: Import libthermal from git.linaro.org/power/libthermal
The thermal library allows to connect the userspace with the thermal
framework. It provides an API to be used in order to get the thermal
topology, the notifications and the setup.

It provides an abstraction of the thermal netlink protocol and it is
event based. Trip point crossed the way up or down are notified via
the netlink socket as well as other notifications like the thermal
zone creation, destruction, enabling or disabling. Furthermore if a
trip point is changed a notification is also delivered with the new
information.

The https://git.linaro.org/power/libthermal repository is itself a
copy of the Linux kernel tools/thermal/libthermal.

The license of this library is LPGL. The library stays
self-encapsulated, so no code inside the dragonboard project is
supposed to copy or use code from this library except using the header
as defined by the LGPL license and link with the shared library
binary.

Please note the author of the library in the same as the author of
this change.

Link: https://git.linaro.org/power/libthermal
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: Id9e1b43a4bf36346257fb5a0ca5031d17434eeeb