device-common: Do not enable UFS in builds booting from sdcard
To simplify sdcard based testing in lab, we do not ship UFS
driver module in vendor_dlkm.img anymore. This will help in
the LKFT testing where we only need to update boot.img now
and not touch userspace images anymore.

The downside is that if we are not shipping UFS driver as a
vendor_dlkm kernel module then we can not ship it in our
build at all because if we ship UFS driver in ramdisk then
the AOSP's first_stage_init will probe UFS first and try to
mount it as the boot block device, irrespective of the
androidboot.boot_devices bootarg value.

Fixes: dc57023acdbd ("db845c: Enable booting from mmc-sdcard")
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: I3002025392eb31c84459e09568944d76f20d34c8