Revert "sepolicy/set_udc.te: fix build error on sysfs_udc type"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb35fd6e-4298-49e5-b90c-3e02f53b0069). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55640ae41ff0af56dda6c45e871e25f076092619
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com