Revert "Switch to android16-6.12 as the default kernel version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f2d21f5-f38f-4d1c-9158-5d0bfe94261d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 387949028
Change-Id: I17c355a34d71ef4f59b025865e148ae98bc5b383
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com