Revert "Switch to android16-6.12 as the default kernel version"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/084c1b9f-57c5-4d90-95ce-31252c1e3d37). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 387942206
Bug: 387949028
Change-Id: I9561515ad752b3bf17eebee8eb4c0a055a57c22f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com